Spherical Harmonic Coefficients

Spherical Harmonic Coefficients

LibHealpix.Alm — Type.

struct Alm{T<:Number} <: AbstractVector{T}

This type holds a vector of spherical harmonic coefficients.

Fields:

lmax - the maximum value for the $l$ quantum number

mmax - the maximum value for the $m$ quantum number (note that $m ≤ l$)

coefficients - the list of spherical harmonic coefficients

Constructors:

Alm(T, lmax, mmax)

Construct an Alm object that will store all spherical harmonic coefficients with element type T, $l ≤ lₘₐₓ$, and $m ≤ mₘₐₓ$. All of the coefficients will be initialized to zero.

Alm(lmax, mmax, coefficients)

Construct an Alm object with the given list of initial coefficients corresponding to $l ≤ lₘₐₓ$, and $m ≤ mₘₐₓ$. A LibHealpixException will be thrown if too many or too few coefficients are provided.

Usage:

julia> alm = Alm(Complex128, 10, 10)
       for (l, m) in lm(alm)
           @lm alm[l, m] = l * m
       end
       @lm(alm[10, 5]) == 50
true

Note

The lm function is used to iterate over the spherical harmonic quantum numbers $l$ and $m$.

Note

The @lm macro is used to index into an Alm object when given the spherical harmonic quantum numbers $l$ and $m$.

See also: RingHealpixMap, NestHealpixMap, lm, @lm

source

LibHealpix.lm — Function.

lm(lmax, mmax)
lm(alm)

Construct an interator for iterating over all possible values of the spherical harmonic quantum numbers $l ≤ lₘₐₓ$ and $m ≤ mₘₐₓ$.

Arguments:

lmax - the maximum value of $l$

mmax - the maximum value of $m$

alm - if an Alm object is provided, lmax and mmax will be inferred from the corresponding fields

Usage:

julia> for (l, m) in lm(2, 1)
           @show l, m
       end
(l, m) = (0, 0)
(l, m) = (1, 0)
(l, m) = (2, 0)
(l, m) = (1, 1)
(l, m) = (2, 1)

See Also: Alm, @lm

source

LibHealpix.@lm — Macro.

@lm

This macro is used to index an Alm object when given the values for quantum numbers $l$ and $m$.

Usage

julia> alm = Alm(Int, 2, 1)
       for (l, m) in lm(alm)
           @lm alm[l, m] = l + m
       end

julia> @lm alm[1, 1]
2

julia> @lm alm[1, :] # all coefficients with l == 1
2-element Array{Int64,1}:
 1
 2

julia> @lm alm[:, 1] # all coefficients with m == 1
2-element Array{Int64,1}:
 2
 3

Background

Alm implements the AbstractVector interface which allows the type to be used in place of a standard Vector in many cases. This generally makes sense because Alm is simply a wrapper around a standard Vector.

However, one consequence of being an AbstractVector is that the two-element getindex function already has a meaning and therefore alm[l, m] cannot be used to mean "give me the coefficient corresponding to the quantum numbers $l$ and $m$". Instead @lm alm[l, m] calls a separate function that does give you the coefficient for $l$ and $m$.

See Also: Alm, lm

source