Healpix Maps

Healpix Maps

LibHealpix.HealpixMap — Type.

abstract type HealpixMap{T<:Number} <: AbstractVector{T}

This abstract type represents a Healpix equal-area pixelization of the sphere.

Subtypes:

RingHealpixMap - a HealpixMap where pixels are ordered along rings of constant latitude. This ordering should be used for performing spherical harmonic transforms.

NestHealpixMap - a HealpixMap where nearby pixels also tend to be nearby in memory.

source

LibHealpix.RingHealpixMap — Type.

struct RingHealpixMap{T<:Number} <: HealpixMap{T}

This type represents a Healpix equal-area pixelization of the sphere where pixels are ordered along rings of constant latitude.

Fields:

nside - the Healpix resolution parameter

pixels - the list of pixel values

Constructors:

RingHealpixMap(T, nside)

Construct a RingHealpixMap with the element type T and resolution parameter nside. All of the pixels will be set to zero initially.

RingHealpixMap(pixels)

Construct a RingHealpixMap with the given list of pixel values. The resolution parameter nside will be inferred from the number of pixels. However a LibHealpixException will be thrown if given an invalid number of pixels.

RingHealpixMap(nside, pixels)

Construct a RingHealpixMap with the given resolution parameter nside and initial list of pixel values. This constructor is cheaper than RingHealpixMap(pixels) if the correct value of nside is already known.

Usage:

julia> map = RingHealpixMap(Float64, 256)
       for idx = 1:length(map)
           map[idx] = randn()
       end
       map + map == 2map
true

See also: HealpixMap, NestHealpixMap, Alm

source

LibHealpix.NestHealpixMap — Type.

struct NestHealpixMap{T<:Number} <: HealpixMap{T}

This type represents a Healpix equal-area pixelization of the sphere where nearby pixels also tend to be nearby in memory.

Fields:

nside - the Healpix resolution parameter

pixels - the list of pixel values

Constructors:

NestHealpixMap(T, nside)

Construct a NestHealpixMap with the element type T and resolution parameter nside. All of the pixels will be set to zero initially.

NestHealpixMap(pixels)

Construct a NestHealpixMap with the given list of pixel values. The resolution parameter nside will be inferred from the number of pixels. However a LibHealpixException will be thrown if given an invalid number of pixels.

NestHealpixMap(nside, pixels)

Construct a NestHealpixMap with the given resolution parameter nside and initial list of pixel values. This constructor is cheaper than NestHealpixMap(pixels) if the correct value of nside is already known.

Usage:

julia> map = NestHealpixMap(Float64, 256)
       for idx = 1:length(map)
           map[idx] = randn()
       end
       map + map == 2map
true

See also: HealpixMap, RingHealpixMap, Alm

source

LibHealpix.ang2pix — Function.

ang2pix(map, theta, phi)

Compute the pixel index that contains the point on the sphere given by the spherical coordinates $(θ, ϕ)$.

Arguments:

map - the input Healpix map

theta - the inclination angle $θ$

phi - the azimuthal angle $ϕ$

Usage:

julia> ang2pix(RingHealpixMap(Float64, 256), π/2, π/2)
392961

julia> ang2pix(NestHealpixMap(Float64, 256), π/2, π/2)
354987

See Also: pix2ang, ang2pix_nest, ang2pix_ring

source

LibHealpix.pix2ang — Function.

pix2ang(map, ipix)

Compute the spherical coordinates $(θ, ϕ)$ corresponding to the given pixel center.

Arguments:

map - the input Healpix map

ipix - the pixel index

Usage:

julia> pix2ang(RingHealpixMap(Float64, 256), 1)
(0.0031894411211228764, 0.7853981633974483)

julia> pix2ang(NestHealpixMap(Float64, 256), 1)
(1.5681921571847817, 0.7853981633974483)

See Also: ang2pix, pix2ang_nest, pix2ang_ring

source

LibHealpix.vec2pix — Function.

vec2pix(map, vec)

Compute the pixel index that contains the point on the sphere given by the Cartesian unit vector.

Arguments:

map - the input Healpix map

vec - the input Cartesian unit vector

Usage:

julia> vec2pix(RingHealpixMap(Float64, 256), [0, 0, 1])
1

julia> vec2pix(NestHealpixMap(Float64, 256), [0, 0, 1])
65536

See Also: pix2vec, vec2pix_nest, vec2pix_ring

source

LibHealpix.pix2vec — Function.

pix2vec(map, ipix)

Compute the Cartesian unit vector corresponding to the given pixel center.

Arguments:

map - the input Healpix map

ipix - the pixel index (nested scheme)

Usage:

julia> pix2vec(RingHealpixMap(Float64, 256), 1)
3-element LibHealpix.UnitVector:
 0.00225527
 0.00225527
 0.999995

julia> pix2vec(NestHealpixMap(Float64, 256), 1)
3-element LibHealpix.UnitVector:
 0.707104
 0.707104
 0.00260417

See Also: vec2pix, pix2vec_nest, pix2vec_ring

source