Pixel Functions

Pixel Functions

LibHealpix.nside2npix — Function.

nside2npix(nside)

Compute the number of pixels in a Healpix map with the given value of nside.

Arguments:

nside - the Healpix resolution parameter

Usage:

julia> nside2npix(4)
192

julia> nside2npix(256)
786432

See Also: npix2nside, nside2nring

source

LibHealpix.npix2nside — Function.

npix2nside(npix)

Compute the value of the nside parameter for a Healpix map with the given number of pixels.

Arguments:

npix - the number of pixels in the map

Usage:

julia> npix2nside(192)
4

julia> npix2nside(786432)
256

See Also: nside2npix, nside2nring

source

LibHealpix.nside2nring — Function.

nside2nring(nside)

Compute the number of equal latitude rings in the Healpix map with the given value of nside.

Arguments:

nside - the Healpix resolution parameter

Usage:

julia> nside2nring(4)
15

julia> nside2nring(256)
1023

See Also: nside2npix, npix2nside

source

LibHealpix.ang2vec — Function.

ang2vec(theta, phi)

Compute the Cartesian unit vector to the spherical coordinates $(θ, ϕ)$.

Arguments:

theta - the inclination angle $θ$

phi - the azimuthal angle $ϕ$

Usage:

julia> ang2vec(0, 0)
3-element LibHealpix.UnitVector:
 0.0
 0.0
 1.0

julia> ang2vec(π/2, π/2)
3-element LibHealpix.UnitVector:
 6.12323e-17
 1.0
 6.12323e-17

See Also: vec2ang

source

LibHealpix.vec2ang — Function.

vec2ang(vec)

Compute the spherical coordinates $(θ, ϕ)$ from the given unit vector.

Arguments:

vec - the input Cartesian unit vector

Usage:

julia> vec2ang([1, 0, 0])
(1.5707963267948966, 0.0)

julia> vec2ang([0, 1, 0])
(1.5707963267948966, 1.5707963267948966)

julia> vec2ang([0, 0, 1])
(0.0, 0.0)

See Also: ang2vec

source

LibHealpix.nest2ring — Function.

nest2ring(nside, ipix)

Convert the given pixel index from the nested to the ring indexing scheme.

Arguments:

nside - the Healpix resolution parameter

ipix - the pixel index (nested scheme)

Usage:

julia> nest2ring(256, 1)
391809

julia> nest2ring(256, 2)
390785

See Also: ring2nest

source

LibHealpix.ring2nest — Function.

ring2nest(nside, ipix)

Convert the given pixel index from the ring to the nested indexing scheme.

Arguments:

nside - the Healpix resolution parameter

ipix - the pixel index (ring scheme)

Usage:

julia> ring2nest(256, 1)
65536

julia> ring2nest(256, 2)
131072

See Also: nest2ring

source

LibHealpix.ang2pix_nest — Function.

ang2pix_nest(nside, theta, phi)

Compute the pixel index (in the nested scheme) that contains the point on the sphere given by the spherical coordinates $(θ, ϕ)$.

Arguments:

nside - the Healpix resolution parameter

theta - the inclination angle $θ$

phi - the azimuthal angle $ϕ$

Usage:

julia> ang2pix_nest(256, 0, 0)
65536

julia> ang2pix_nest(256, π/2, π/2)
354987

See Also: ang2pix_ring, pix2ang_nest, pix2ang_ring

source

LibHealpix.ang2pix_ring — Function.

ang2pix_ring(nside, theta, phi)

Compute the pixel index (in the ring scheme) that contains the point on the sphere given by the spherical coordinates $(θ, ϕ)$.

Arguments:

nside - the Healpix resolution parameter

theta - the inclination angle $θ$

phi - the azimuthal angle $ϕ$

Usage:

julia> ang2pix_ring(256, 0, 0)
1

julia> ang2pix_ring(256, π/2, π/2)
392961

See Also: ang2pix_nest, pix2ang_nest, pix2ang_ring

source

LibHealpix.pix2ang_nest — Function.

pix2ang_nest(nside, ipix)

Compute the spherical coordinates $(θ, ϕ)$ corresponding to the given pixel center.

Arguments:

nside - the Healpix resolution parameter

ipix - the pixel index (nested scheme)

Usage:

julia> pix2ang_nest(256, 1)
(1.5681921571847817, 0.7853981633974483)

julia> pix2ang_nest(256, 2)
(1.5655879699137618, 0.7884661249732196)

See Also: pix2ang_ring, ang2pix_nest, ang2pix_ring

source

LibHealpix.pix2ang_ring — Function.

pix2ang_ring(nside, ipix)

Compute the spherical coordinates $(θ, ϕ)$ corresponding to the given pixel center.

Arguments:

nside - the Healpix resolution parameter

ipix - the pixel index (ring scheme)

Usage:

julia> pix2ang_ring(256, 1)
(0.0031894411211228764, 0.7853981633974483)

julia> pix2ang_ring(256, 2)
(0.0031894411211228764, 2.356194490192345)

See Also: pix2ang_nest, ang2pix_nest, ang2pix_ring

source

LibHealpix.vec2pix_nest — Function.

vec2pix_nest(nside, vec)

Compute the pixel index (in the nested scheme) that contains the point on the sphere given by the Cartesian unit vector.

Arguments:

nside - the Healpix resolution parameter

vec - the input Cartesian unit vector

Usage:

julia> vec2pix_nest(256, [1, 0, 0])
289451

julia> vec2pix_nest(256, [0, 1, 0])
354987

julia> vec2pix_nest(256, [0, 0, 1])
65536

See Also: vec2pix_ring, pix2vec_nest, pix2vec_ring

source

LibHealpix.vec2pix_ring — Function.

vec2pix_ring(nside, vec)

Compute the pixel index (in the ring scheme) that contains the point on the sphere given by the Cartesian unit vector.

Arguments:

nside - the Healpix resolution parameter

vec - the input Cartesian unit vector

Usage:

julia> vec2pix_ring(256, [1, 0, 0])
392705

julia> vec2pix_ring(256, [0, 1, 0])
392961

julia> vec2pix_ring(256, [0, 0, 1])
1

See Also: vec2pix_nest, pix2vec_nest, pix2vec_ring

source

LibHealpix.pix2vec_nest — Function.

pix2vec_nest(nside, ipix)

Compute the Cartesian unit vector corresponding to the given pixel center.

Arguments:

nside - the Healpix resolution parameter

ipix - the pixel index (nested scheme)

Usage:

julia> pix2vec_nest(256, 1)
3-element LibHealpix.UnitVector:
 0.707104
 0.707104
 0.00260417

julia> pix2vec_nest(256, 2)
3-element LibHealpix.UnitVector:
 0.704925
 0.709263
 0.00520833

See Also: pix2vec_ring, vec2pix_nest, vec2pix_ring

source

LibHealpix.pix2vec_ring — Function.

pix2vec_ring(nside, ipix)

Compute the Cartesian unit vector corresponding to the given pixel center.

Arguments:

nside - the Healpix resolution parameter

ipix - the pixel index (ring scheme)

Usage:

julia> pix2vec_ring(256, 1)
3-element LibHealpix.UnitVector:
 0.00225527
 0.00225527
 0.999995

julia> pix2vec_ring(256, 2)
3-element LibHealpix.UnitVector:
 -0.00225527
  0.00225527
  0.999995

See Also: pix2vec_nest, vec2pix_nest, vec2pix_ring

source